+++ native/c/test/zo.system.test.cpp
@@ -37,6 +37,18 @@ void zo_system_tests()
                   ExpectWithContext(rc, response).ToBeGreaterThanOrEqualTo(0);
                 });
            });
+  describe("list-parmlib tests",
+           [&]() -> void
+           {
+             it("should list parmlib",
+                []()
+                {
+                  int rc = 0;
+                  string response;
+                  rc = execute_command_with_output(zowex_command + " system list-parmlib", response);
+                  ExpectWithContext(rc, response).ToBeGreaterThanOrEqualTo(0);
+                });
+           });
   describe("list-apf tests",
            [&]() -> void
            {